+++ docker-compose.yml
@@ -48,18 +48,6 @@ services:
             - '${FORWARD_MAILHOG_DASHBOARD_PORT:-8025}:8025'
         networks:
             - sail
-    redis:
-        image: 'redis:alpine'
-        ports:
-            - '${FORWARD_REDIS_PORT:-6379}:6379'
-        volumes:
-            - 'sailredis:/data'
-        networks:
-            - sail
-        healthcheck:
-            test: ["CMD", "redis-cli", "ping"]
-            retries: 3
-            timeout: 5s
     soketi:
         image: 'quay.io/soketi/soketi:latest-16-alpine'
         environment:
@@ -68,6 +56,26 @@ services:
         ports:
             - '${SOKETI_PORT:-6001}:6001'
             - '${SOKETI_METRICS_SERVER_PORT:-9601}:9601'
+        networks:
+            - sail
+    nodejs:
+        image: 'node:16-alpine'
+        user: 'node'
+        environment:
+            - NODE_ENV=production
+        volumes:
+            - './mqtt_http_bridge:/home/node/mqtt_http_bridge'
+        working_dir: '/home/node/mqtt_http_bridge'
+        command: 'node index.js'
+        networks:
+            - sail
+    mosquitto:
+        image: 'eclipse-mosquitto:latest'
+        volumes:
+          - './mosquitto.conf:/mosquitto/config/mosquitto.conf'
+        ports:
+          - 1883:1883
+          - 9001:9001
         networks:
             - sail
 networks:
